refactor(sonar-idl): clean up and improve maintainability
Breaking changes to sonar-idl crate:

- Split models.rs into focused modules (core, account, types, legacy)
- Replace OrderedJsonValue with serde_json::Value for ecosystem compatibility
- Add typed errors with thiserror (IdlError enum)
- Unify lookup: IdlLookup trait only for ResolvedIdl, not Idl
- Add Idl::parse() constructor for one-step IDL loading
- Consolidate discriminator constants (NAMESPACE_*, CPI_EVENT_DISCRIMINATOR)
- Add Discriminator struct with from_bytes()/as_bytes() methods
- Remove IdlTypeDefinitionKind::Other variant
- Move integration tests to tests/ directory

Dependencies:
- Remove anyhow
- Add thiserror, indexmap
