+++ crates/sonar-idl/Cargo.toml
@@ -6,9 +6,10 @@ rust-version.workspace = true
 description = "Anchor IDL parsing and type resolution for Solana programs"
 
 [dependencies]
-anyhow = "1.0"
 hex = "0.4"
+indexmap = { version = "2.0", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 sha2 = "0.10.9"
 solana-pubkey = "3.0"
+thiserror = "1.0"
